break accordion out into seperate components
